requirements.txt
torch==2.0.1
torchvision==0.15.2
pandas==2.0.3
matplotlib==3.7.3
PyYAML==6.0.1
wandb==0.15.10
albumentations==1.3.1
tensorboard==2.14.0
scikit_learn==1.3.0
black